refactor: moved atclient crud implementations to their individual files
