+++ examples/desktop/at_talk/src/main.c
@@ -76,14 +76,14 @@ int main(int argc, char **argv) {
   // Init variables and get the encryption keys
 
   char *enc_key_shared_by_me = malloc(45);
-  ret = atclient_get_encryption_key_shared_by_me(&atclient, &recipient, enc_key_shared_by_me, true);
+  ret = atclient_get_shared_encryption_key_shared_by_me(&atclient, &recipient, enc_key_shared_by_me, true);
   if (ret != 0) {
     free(enc_key_shared_by_me);
     goto exit2;
   }
 
   char *enc_key_shared_by_other = malloc(45);
-  ret = atclient_get_encryption_key_shared_by_other(&atclient, &recipient, enc_key_shared_by_other);
+  ret = atclient_get_shared_encryption_key_shared_by_other(&atclient, &recipient, enc_key_shared_by_other);
   if (ret != 0) {
     free(enc_key_shared_by_me);
     free(enc_key_shared_by_other);
