+++ packages/atclient/CMakeLists.txt
@@ -6,7 +6,6 @@ set(ATCLIENT_INCLUDE_DIR ${CMAKE_CURRENT_LIST_DIR}/include)
 set(
   ATCLIENT_SOURCES
   ${CMAKE_CURRENT_LIST_DIR}/src/atbytes.c
-  ${CMAKE_CURRENT_LIST_DIR}/src/atclient.c
   ${CMAKE_CURRENT_LIST_DIR}/src/atkey.c
   ${CMAKE_CURRENT_LIST_DIR}/src/atkeys.c
   ${CMAKE_CURRENT_LIST_DIR}/src/atkeysfile.c
@@ -16,6 +15,14 @@ set(
   ${CMAKE_CURRENT_LIST_DIR}/src/metadata.c
   ${CMAKE_CURRENT_LIST_DIR}/src/notification.c
   ${CMAKE_CURRENT_LIST_DIR}/src/stringutils.c
+  ${CMAKE_CURRENT_LIST_DIR}/src/atclient_delete.c
+  ${CMAKE_CURRENT_LIST_DIR}/src/atclient_get_publickey.c
+  ${CMAKE_CURRENT_LIST_DIR}/src/atclient_get_selfkey.c
+  ${CMAKE_CURRENT_LIST_DIR}/src/atclient_get_sharedkey.c
+  ${CMAKE_CURRENT_LIST_DIR}/src/atclient_put.c
+  ${CMAKE_CURRENT_LIST_DIR}/src/encryption_key_helpers.c
+  ${CMAKE_CURRENT_LIST_DIR}/src/encryption_key_helpers.c
+  ${CMAKE_CURRENT_LIST_DIR}/src/atclient.c
 )
 
 # Project setup
