+++ packages/atclient/include/atclient/encryption_key_helpers.h
@@ -1,0 +1,78 @@
+#include "atclient/atclient.h"
+#include "atclient/atkey.h"
+#include "atclient/atkeys.h"
+#include "atlogger/atlogger.h"
+
+/**
+ * @brief Looks up the symmetric shared key which the atclient's atsign shared with the recipient's atsign.
+ * If no key is found and create_new_if_not_found is true, it will create, store and share a new one with the
+ * recipient's atsign.
+ *
+ * @param ctx Initialized atclient context (required)
+ * @param recipient An atclient_atsign struct corresponding to the atsign with whom the key was shared (required)
+ * @param enc_key_shared_by_me The output shared key in b64 format (required)
+ * @param create_new_if_not_found true if in case the symmetric shared key does not exist, you would like it to be
+ * created / false if not (required)
+ * @return int 0 on success, error otherwise
+ */
+int atclient_get_shared_encryption_key_shared_by_me(atclient *ctx, const atclient_atsign *recipient,
+                                                    char *enc_key_shared_by_me, bool create_new_if_not_found);
+
+/**
+ * @brief Looks up the symmetric shared key which the recipient's atsign shared with atclient's atsign.
+ * If no key is found, the function will return an error.
+ *
+ * @param ctx Initialized atclient context (required)
+ * @param root_conn initialized root connection
+ * @param recipient An atclient_atsign struct corresponding to the atsign who shared the key with the atclient’s atsign
+ * (required)
+ * @param enc_key_shared_by_other the output shared key in b64 format (required)
+ * @return int 0 on success, error otherwise
+ */
+int atclient_get_shared_encryption_key_shared_by_other(atclient *ctx, const atclient_atsign *recipient,
+                                                       char *enc_key_shared_by_other);
+
+/**
+ * @brief Creates a symmetric shared key, which the atclient atsign shares with the recipient atsign.
+ *
+ * @param ctx Initialized atclient context (required)
+ * @param root_conn initialized root connection
+ * @param recipient An atclient_atsign struct corresponding to the atsign with which you want to create the shared key
+ * (required)
+ * @param enc_key_shared_by_me The output new shared key (which was already stored in the server) in b64 format
+ * (required)
+ * @return int 0 on success, error otherwise
+ */
+int atclient_create_shared_encryption_key(atclient *ctx, atclient_connection *root_conn,
+                                          const atclient_atsign *recipient, char *enc_key_shared_by_me);
+
+/**
+ * @brief Retreives the public encryption key of a given atsign.
+ *
+ * @param ctx Initialized atclient context (required)
+ * @param root_conn initialized root connection
+ * @param recipient An atclient_atsign struct corresponding to the atsign which public encryption key you would like to
+ * obtain. It may receive a NULL value, in which case, the atclient_atsign contained in the ctx parameter will be used
+ * (required)
+ * @param public_encryption_key The output public key in b64 format (required)
+ * @return int 0 on success, error otherwise
+ */
+int atclient_get_public_encryption_key(atclient *ctx, atclient_connection *root_conn, const atclient_atsign *atsign,
+                                       char *public_encryption_key);
+
+/**
+ * @brief Creates a shared encryption key pair and puts it in your atServer. One is made for you to use for encrypting
+ * data to send to other, and the other is for the recipient to use to decrypt the data that you sent
+ * (shared_key.other@me and @other:shared_key@me)
+ *
+ * @param atclient the atclient context (must be initialized and pkam_authenticated)
+ * @param root_conn the root connection (must be initialized)
+ * @param sharedby 
+ * @param sharedwith
+ * @param sharedenckeybyme
+ * @return int
+ */
+int atclient_create_shared_encryption_key_pair_for_me_and_other(atclient *atclient, atclient_connection *root_conn,
+                                                                const atclient_atsign *sharedby,
+                                                                const atclient_atsign *sharedwith,
+                                                                char *sharedenckeybyme);
